# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/nutri/foods": {
      "get": {
        "description": "Busca fuzzy: insensível a maiúscula E a acento (\"acai\" acha \"Açaí\"), e por subsequência — \"arrint\" acha \"Arroz integral\". O resultado vem ordenado por relevância (casamento colado e em início de palavra na frente). `q` vazio devolve a primeira página, não erro — é o estado inicial da tela. `total` diz quantos casaram; a lista é a fatia `[offset, offset+limit)`, então há mais páginas enquanto `offset + foods.length < total`. `limit`/`offset` fora de forma caem no default em vez de 400.",
        "operationId": "CatalogoController_buscarAlimentos",
        "parameters": [
          {
            "name": "x-nutri-key",
            "in": "header",
            "description": "credencial stub da nutri (env NUTRI_API_KEY; fail-closed)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "trecho do nome",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "1–600, default 50",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "quantos pular, default 0",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "alimentos que casaram + total"
          },
          "403": {
            "description": "x-nutri-key ausente/errada"
          }
        },
        "summary": "Buscar alimentos",
        "tags": [
          "Catálogo do editor (só nutri)"
        ]
      },
      "post": {
        "description": "Exige nome e os quatro nutrientes por 100 g (zero é válido: água tem 0 kcal). Nasce SEM `taco_id` e com `source` diferente de \"taco\", então a ingestão TACO (008), que faz upsert por `taco_id`, nunca o sobrescreve.",
        "operationId": "CatalogoController_criarAlimento",
        "parameters": [
          {
            "name": "x-nutri-key",
            "in": "header",
            "description": "credencial stub da nutri (env NUTRI_API_KEY; fail-closed)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "example": {
                  "name": "Pão da padaria da esquina",
                  "kcalPer100g": 270,
                  "carbPer100g": 55,
                  "proteinPer100g": 8,
                  "fatPer100g": 2,
                  "fiberPer100g": 2.5
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "o alimento criado"
          },
          "400": {
            "description": "nome ou nutriente inválido"
          }
        },
        "summary": "Cadastrar um alimento fora da TACO",
        "tags": [
          "Catálogo do editor (só nutri)"
        ]
      }
    },
    "/nutri/foods/{foodId}": {
      "patch": {
        "description": "PATCH parcial. Vale para alimento da TACO também — mas a próxima ingestão sobrescreve o que tiver `taco_id`.",
        "operationId": "CatalogoController_atualizarAlimento",
        "parameters": [
          {
            "name": "x-nutri-key",
            "in": "header",
            "description": "credencial stub da nutri (env NUTRI_API_KEY; fail-closed)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "foodId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "o alimento depois da alteração"
          },
          "404": {
            "description": "alimento não encontrado"
          }
        },
        "summary": "Editar um alimento",
        "tags": [
          "Catálogo do editor (só nutri)"
        ]
      },
      "delete": {
        "description": "Leva os vínculos de grupo e as medidas caseiras. RECUSA com 409 se o alimento está em algum plano ou em algum registro de refeição.",
        "operationId": "CatalogoController_excluirAlimento",
        "parameters": [
          {
            "name": "x-nutri-key",
            "in": "header",
            "description": "credencial stub da nutri (env NUTRI_API_KEY; fail-closed)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "foodId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "alimento excluído"
          },
          "404": {
            "description": "alimento não encontrado"
          },
          "409": {
            "description": "em uso por plano ou registro"
          }
        },
        "summary": "Excluir um alimento do catálogo",
        "tags": [
          "Catálogo do editor (só nutri)"
        ]
      }
    },
    "/nutri/substitution-groups": {
      "get": {
        "description": "`custom: true` = grupo desta nutri; `false` = grupo do sistema, mantido pela auto-classificação (008). Cada vínculo traz a porção de referência (a \"1 troca\") e a origem (manual = curadoria, nunca sobrescrita).",
        "operationId": "CatalogoController_listarGrupos",
        "parameters": [
          {
            "name": "x-nutri-key",
            "in": "header",
            "description": "credencial stub da nutri (env NUTRI_API_KEY; fail-closed)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "grupos + vínculos"
          }
        },
        "summary": "Listar os grupos de substituição com os alimentos vinculados",
        "tags": [
          "Catálogo do editor (só nutri)"
        ]
      },
      "post": {
        "description": "`basis` é o nutriente PRESERVADO na troca dentro do grupo: carb | protein | fat | kcal. O grupo nasce vinculado à nutricionista (custom), não ao sistema.",
        "operationId": "CatalogoController_criarGrupo",
        "parameters": [
          {
            "name": "x-nutri-key",
            "in": "header",
            "description": "credencial stub da nutri (env NUTRI_API_KEY; fail-closed)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "example": {
                  "name": "Meus carboidratos",
                  "basis": "carb"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "o grupo criado, sem alimentos"
          },
          "400": {
            "description": "name vazio ou basis fora do conjunto"
          },
          "422": {
            "description": "nenhuma nutricionista cadastrada, ou mais de uma"
          }
        },
        "summary": "Criar um grupo de substituição próprio",
        "tags": [
          "Catálogo do editor (só nutri)"
        ]
      }
    },
    "/nutri/substitution-groups/{groupId}": {
      "patch": {
        "description": "Atenção: mudar `basis` muda o nutriente preservado em TODAS as trocas dentro do grupo.",
        "operationId": "CatalogoController_atualizarGrupo",
        "parameters": [
          {
            "name": "x-nutri-key",
            "in": "header",
            "description": "credencial stub da nutri (env NUTRI_API_KEY; fail-closed)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "o grupo depois da alteração"
          },
          "404": {
            "description": "grupo não encontrado"
          }
        },
        "summary": "Editar o grupo",
        "tags": [
          "Catálogo do editor (só nutri)"
        ]
      },
      "delete": {
        "description": "RECUSA com 409 se algum item de plano é flexível dentro dele.",
        "operationId": "CatalogoController_excluirGrupo",
        "parameters": [
          {
            "name": "x-nutri-key",
            "in": "header",
            "description": "credencial stub da nutri (env NUTRI_API_KEY; fail-closed)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "grupo excluído"
          },
          "404": {
            "description": "grupo não encontrado"
          },
          "409": {
            "description": "em uso por item de plano"
          }
        },
        "summary": "Excluir o grupo (com os vínculos)",
        "tags": [
          "Catálogo do editor (só nutri)"
        ]
      }
    },
    "/nutri/substitution-groups/{groupId}/foods/{foodId}": {
      "put": {
        "description": "A porção de referência é a \"1 troca\" do exchange — é dela que sai o recálculo de quantidade na substituição. PUT porque o par (grupo, alimento) é a identidade: revincular atualiza. O vínculo nasce/vira `origin: manual`, que a auto-classificação nunca sobrescreve.",
        "operationId": "CatalogoController_vincular",
        "parameters": [
          {
            "name": "x-nutri-key",
            "in": "header",
            "description": "credencial stub da nutri (env NUTRI_API_KEY; fail-closed)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "foodId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "example": {
                  "referencePortionGrams": 100
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "o grupo com o vínculo"
          },
          "400": {
            "description": "referencePortionGrams ≤ 0"
          },
          "404": {
            "description": "grupo ou alimento não encontrado"
          }
        },
        "summary": "Vincular alimento ao grupo com a porção de referência",
        "tags": [
          "Catálogo do editor (só nutri)"
        ]
      },
      "delete": {
        "description": "RECUSA com 409 se existe item de plano deste alimento marcado como flexível dentro deste grupo — sem o vínculo a troca perderia a porção de referência.",
        "operationId": "CatalogoController_desvincular",
        "parameters": [
          {
            "name": "x-nutri-key",
            "in": "header",
            "description": "credencial stub da nutri (env NUTRI_API_KEY; fail-closed)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "foodId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "o grupo sem o vínculo"
          },
          "404": {
            "description": "grupo não encontrado"
          },
          "409": {
            "description": "item de plano depende do vínculo"
          }
        },
        "summary": "Desvincular alimento do grupo",
        "tags": [
          "Catálogo do editor (só nutri)"
        ]
      }
    }
  },
  "info": {
    "title": "Bamboo API",
    "description": "API da alça do paciente — ver \"o agora\" (US1) e substituir um alimento dentro do grupo, com quantidade recalculada e medida caseira (US2). v0: auth stub (paciente fixo por env).",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
...
}